repo hirerchy and improve storybook
