+++ app/src/organisms/Desktop/Devices/RobotSettings/RobotSettingsComplianceReady/userAccount/UserAccountConfirmModal/useraccountconfirmmodal.module.css
@@ -1,0 +1,25 @@
+.content {
+  display: flex;
+  padding: var(--spacing-24);
+  flex-direction: column;
+  align-items: flex-start;
+  gap: var(--spacing-24);
+  align-self: stretch;
+}
+
+.body {
+  display: flex;
+  width: 100%;
+  flex-direction: column;
+  align-items: center;
+  align-self: stretch;
+  text-align: center;
+  gap: var(--spacing-8);
+}
+
+.actions {
+  display: flex;
+  width: 100%;
+  justify-content: flex-end;
+  gap: var(--spacing-8);
+}
